Run Rustfmt, resolve Clippy lints, use 2018 edition (#31)
I do Rust as part of my living, and I wanted to update this codebase to 
match with the standard way Rust repositories are managed. This PR is 
mostly running Rustfmt and fixing Clippy errors. These are both 
invaluable tools I highly recommend you try out if you haven't already. 
Rustfmt is an autoformatter that not only makes it so there's no 
conflicting styles, but also makes it really easy to do big refactors 
without having to care about style too much. VS Code has an option to 
run a formatter when you save, and if you're using a Rust LSP (RLS or 
the newer rust-analyzer), then it helps productivity a lot. Clippy is a 
very good linter that finds issues that the normal rustc doesn't catch. 
For this repo, it found some style issues and a minor performance issue 
that were easily fixed up.

A lot of the code is inside macros, which clippy and Rustfmt can't 
introspect very well, but this covers most of it.

- Runs cargo fmt*
- Fixes error brought up by cargo clippy*
- Upgrades to Rust 2018, which is regularly updated
- Change `const _: () = "bla"` to `compile_error!("bla")`
- Change `use` to be use the idiomatic grouped syntax

* - I highly recommend setting up a GitHub action to do these.