Use the full path rather than filename only in logging cache
Apparently there are some circumstances on Linux where the cache is not
cleared between rounds. Although it would be good to ensure the cache is
cleared, at a minimum there's no reason for rust-g not to support
writing to both "one/foo.txt" and "two/foo.txt" in the same session.