Add an a* pathfinder (#113)
Returns the shortest path in a static node map. That is made for TGMC wich uses manually placed nodes for pathfinding.

Benchmark :
image
That's the average number of path computed in one second, out of 30 runs with random nodes. Size of the node map is roughly 800

On average 32 times faster than tgmc a* implementation.
Another possible comparison is with TG's JPS system, and according to the benchmark done here it's 7000 times faster (tgstation/tgstation#56780). Not exactly the same use cases though